+++ src/main/java/luamade/luawrap/LuaInterface.java
@@ -1,4 +1,4 @@
-package luamade.lua;
+package luamade.luawrap;
 
 import org.luaj.vm2.LuaFunction;
